Move LogsList composable to a common space
