+++ app/src/main/java/com/dp/logcatapp/ui/screens/HomeScreen.kt
@@ -17,18 +17,12 @@ import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.slideInHorizontally
 import androidx.compose.animation.slideOutHorizontally
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.interaction.InteractionSource
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.PressInteraction.Press
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.consumeWindowInsets
@@ -36,15 +30,10 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.text.selection.LocalTextSelectionColors
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ViewList
@@ -72,7 +61,6 @@ import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.IconButtonDefaults
@@ -116,11 +104,7 @@ import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat
@@ -131,7 +115,6 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import com.dp.logcat.Filter
 import com.dp.logcat.Log
-import com.dp.logcat.LogPriority
 import com.dp.logcat.LogcatUtil
 import com.dp.logcatapp.BuildConfig
 import com.dp.logcatapp.R
@@ -147,12 +130,10 @@ import com.dp.logcatapp.services.LogcatService
 import com.dp.logcatapp.services.getService
 import com.dp.logcatapp.ui.common.Dialog
 import com.dp.logcatapp.ui.common.LOGCAT_DIR
-import com.dp.logcatapp.ui.screens.SearchHitKey.LogComponent
+import com.dp.logcatapp.ui.common.LogsList
+import com.dp.logcatapp.ui.common.SearchHitKey
+import com.dp.logcatapp.ui.common.SearchHitKey.LogComponent
 import com.dp.logcatapp.ui.theme.AppTypography
-import com.dp.logcatapp.ui.theme.LogPriorityColors
-import com.dp.logcatapp.ui.theme.LogcatReaderTheme
-import com.dp.logcatapp.ui.theme.RobotoMonoFontFamily
-import com.dp.logcatapp.ui.theme.currentSearchHitColor
 import com.dp.logcatapp.util.PreferenceKeys
 import com.dp.logcatapp.util.ServiceBinder
 import com.dp.logcatapp.util.ShareUtils
@@ -1058,194 +1039,6 @@ private fun TopSearchBar(
   )
 }
 
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun LogsList(
-  modifier: Modifier,
-  contentPadding: PaddingValues,
-  logs: List<Log>,
-  searchHits: Map<SearchHitKey, Pair<Int, Int>>,
-  currentSearchHitLogId: Int,
-  onClick: (Int) -> Unit,
-  onLongClick: (Int) -> Unit,
-  state: LazyListState = rememberLazyListState(),
-) {
-  val textSelectionColors = LocalTextSelectionColors.current
-  val currentSearchHitColor = currentSearchHitColor()
-  LazyColumn(
-    modifier = modifier,
-    state = state,
-    contentPadding = contentPadding,
-  ) {
-    itemsIndexed(
-      items = logs,
-      key = { index, _ -> logs[index].id }
-    ) { index, item ->
-      if (index > 0) {
-        HorizontalDivider()
-      }
-
-      fun maybeHighlightSearchHit(target: String, searchHitKey: SearchHitKey): AnnotatedString {
-        val hit = if (searchHits.isNotEmpty()) searchHits[searchHitKey] else null
-        return if (hit != null) {
-          buildAnnotatedString {
-            append(target)
-            val color = if (index == currentSearchHitLogId) {
-              currentSearchHitColor
-            } else {
-              textSelectionColors.backgroundColor
-            }
-            addStyle(
-              SpanStyle(
-                background = color
-              ),
-              start = hit.first,
-              end = hit.second
-            )
-          }
-        } else {
-          AnnotatedString(target)
-        }
-      }
-
-      LogItem(
-        modifier = Modifier
-          .fillMaxWidth()
-          .combinedClickable(
-            onLongClick = { onLongClick(index) },
-            onClick = { onClick(index) },
-          )
-          .wrapContentHeight(),
-        priority = item.priority,
-        tag = maybeHighlightSearchHit(
-          target = item.tag,
-          searchHitKey = SearchHitKey(logId = item.id, component = LogComponent.TAG),
-        ),
-        message = maybeHighlightSearchHit(
-          target = item.msg,
-          searchHitKey = SearchHitKey(logId = item.id, component = LogComponent.MSG),
-        ),
-        date = item.date,
-        time = item.time,
-        pid = item.pid,
-        tid = item.tid,
-        priorityColor = when (item.priority) {
-          LogPriority.ASSERT -> LogPriorityColors.priorityAssert
-          LogPriority.DEBUG -> LogPriorityColors.priorityDebug
-          LogPriority.ERROR -> LogPriorityColors.priorityError
-          LogPriority.FATAL -> LogPriorityColors.priorityFatal
-          LogPriority.INFO -> LogPriorityColors.priorityInfo
-          LogPriority.VERBOSE -> LogPriorityColors.priorityVerbose
-          LogPriority.WARNING -> LogPriorityColors.priorityWarning
-          else -> LogPriorityColors.prioritySilent
-        },
-      )
-    }
-  }
-}
-
-@Composable
-private fun LogItem(
-  modifier: Modifier,
-  priority: String,
-  tag: AnnotatedString,
-  message: AnnotatedString,
-  date: String,
-  time: String,
-  pid: String,
-  tid: String,
-  priorityColor: Color,
-) {
-  Row(
-    modifier = modifier
-      .height(IntrinsicSize.Max),
-  ) {
-    Box(
-      modifier = Modifier
-        .fillMaxHeight()
-        .background(priorityColor)
-        .padding(5.dp),
-    ) {
-      Text(
-        modifier = Modifier.align(Alignment.Center),
-        text = priority,
-        style = TextStyle.Default.copy(
-          fontSize = 12.sp,
-          fontFamily = RobotoMonoFontFamily,
-          fontWeight = FontWeight.Bold,
-          color = Color.White,
-          textAlign = TextAlign.Center,
-        ),
-      )
-    }
-    Column(
-      modifier = Modifier
-        .fillMaxWidth()
-        .padding(all = 5.dp),
-      verticalArrangement = Arrangement.spacedBy(2.dp),
-    ) {
-      Text(
-        modifier = Modifier.fillMaxWidth(),
-        text = tag,
-        style = TextStyle.Default.copy(
-          fontSize = 13.sp,
-          fontFamily = RobotoMonoFontFamily,
-          fontWeight = FontWeight.Medium,
-        )
-      )
-      Text(
-        modifier = Modifier.fillMaxWidth(),
-        text = message,
-        style = TextStyle.Default.copy(
-          fontSize = 12.sp,
-          fontFamily = RobotoMonoFontFamily,
-        )
-      )
-      Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.Absolute.spacedBy(10.dp),
-      ) {
-        Text(
-          text = date,
-          style = TextStyle.Default.copy(
-            fontSize = 12.sp,
-            fontFamily = RobotoMonoFontFamily,
-            fontWeight = FontWeight.Bold,
-            color = Color.Gray,
-          )
-        )
-        Text(
-          text = time,
-          style = TextStyle.Default.copy(
-            fontSize = 12.sp,
-            fontFamily = RobotoMonoFontFamily,
-            fontWeight = FontWeight.Bold,
-            color = Color.Gray,
-          )
-        )
-        Text(
-          text = pid,
-          style = TextStyle.Default.copy(
-            fontSize = 12.sp,
-            fontFamily = RobotoMonoFontFamily,
-            fontWeight = FontWeight.Bold,
-            color = Color.Gray,
-          )
-        )
-        Text(
-          text = tid,
-          style = TextStyle.Default.copy(
-            fontSize = 12.sp,
-            fontFamily = RobotoMonoFontFamily,
-            fontWeight = FontWeight.Bold,
-            color = Color.Gray,
-          )
-        )
-      }
-    }
-  }
-}
-
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 private fun MaybeShowPermissionRequiredDialog() {
@@ -1633,16 +1426,6 @@ enum class RecordStatus {
   SaveRecordedLogs,
 }
 
-data class SearchHitKey(
-  val logId: Int,
-  val component: LogComponent,
-) {
-  enum class LogComponent {
-    MSG,
-    TAG,
-  }
-}
-
 sealed interface SavedLogsBottomSheetState {
   data object Hide : SavedLogsBottomSheetState
   data class Show(
@@ -1686,24 +1469,4 @@ private fun isReadLogsPermissionGranted(context: Context): Boolean {
     context,
     Manifest.permission.READ_LOGS
   ) == PackageManager.PERMISSION_GRANTED
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun LogItemPreview() {
-  LogcatReaderTheme {
-    LogItem(
-      modifier = Modifier
-        .fillMaxWidth()
-        .wrapContentHeight(),
-      priority = "D",
-      tag = AnnotatedString("Tag"),
-      message = AnnotatedString("This is a log"),
-      date = "01-12",
-      time = "21:10:46.123",
-      pid = "1600",
-      tid = "123123",
-      priorityColor = LogPriorityColors.priorityDebug,
-    )
-  }
 }
